requirements.txt
psycopg2-binary>=2.9
requests>=2.31
lxml>=5.1
PyYAML>=6.0